Add serial connectivity checks and port auto-detect
I was having issues with my PC blue screening, it looks by the crash dumps because Deej was trying to continue using a serial port that was no longer available because Windows being Windows. Have added some basic checks to check for serial connectivity. Probably still a window for the issue to crop up but I'll take one attempted read from a non-existent serial port over a loop constantly polling for it.